implemented pages controller and profile page
