+++ app/controllers/quests_controller.rb
@@ -1,5 +1,4 @@
 class QuestsController < ApplicationController
-  skip_before_action :authenticate_user!, only: :index
 
   def index
     @quests = Quest.all
