+++ app/views/pages/profile.html.erb
@@ -1,0 +1,6 @@
+<h1><%= @user.username %>'s Profile</h1>
+
+<p>Level: <%= @user.level %></p>
+<p>Abilities: <%= @user.abilities %></p>
+<p>Class: <%= @user.character_class %></p>
+<p>Stats: <%= @user.stats %></p>
